+++ src/main/java/com/thf/service/ModuleService.java
@@ -4,8 +4,8 @@ import com.thf.common.oo.ResultVO;
 import com.thf.entity.Module;
 
 public interface ModuleService {
-    ResultVO createModule(String token,int pid ,Module module);
+    ResultVO createModule(String token,long pid ,Module module);
     ResultVO updateModule(String token,Module module);
-    ResultVO deleteModule(String token,int[] midlist);
-    ResultVO searchById(int mid);
+    ResultVO deleteModule(String token,long[] midlist);
+    ResultVO searchById(long mid);
 }
